+++ packages/crypto/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@shelve/utils",
-  "version": "0.0.0",
+  "name": "@shelve/crypto",
+  "version": "1.0.0",
   "type": "module",
   "private": true,
   "scripts": {
